Refactor: Split toolbox javascript into multiple files
